+++ internal/core/openapi/src/main/kotlin/com/recco/internal/core/openapi/model/AppUserRecommendationDTO.kt
@@ -32,6 +32,7 @@ import com.squareup.moshi.JsonClass
  * @param lead 
  * @param imageUrl 
  * @param imageAlt 
+ * @param dynamicImageResizingUrl Responsive image url. Supports transformation via query params. Allowed query params key=values width=number, height=number, quality=1..100, format=auto|jgp|png|webp|tiff, fit=cover|contain|inside|outside
  */
 @JsonClass(generateAdapter = true)
 data class AppUserRecommendationDTO(
@@ -61,5 +62,9 @@ data class AppUserRecommendationDTO(
     val imageUrl: kotlin.String? = null,
 
     @Json(name = "imageAlt")
-    val imageAlt: kotlin.String? = null
+    val imageAlt: kotlin.String? = null,
+
+    /* Responsive image url. Supports transformation via query params. Allowed query params key=values width=number, height=number, quality=1..100, format=auto|jgp|png|webp|tiff, fit=cover|contain|inside|outside */
+    @Json(name = "dynamicImageResizingUrl")
+    val dynamicImageResizingUrl: kotlin.String? = null
 )